Add `groupByCard` function in `Purchase`
This commit introduces a `groupByCard` function in the `Purchase` class. This provides the ability to group a list of payments by credit cards, generating a list of payments after evaluating the results with a custom reduction function. This function increases the code's versatility by providing more ways to organize and process payments data.